Make the utils become more generic.
